Imported Simple Kepler Orbits Asset
downloaded the free Kepler Orbits Asset on the unity store.

I did want to make this myself, but it's a lot of maths, stuff that whilst I am good at, I am too lazy to work through the entire thing. I'll look through the assets to try and deconstruct it and maybe produce a light weight version that is suitable for the game, but for now, Kepler orbits are in!

Link to the asset store: https://assetstore.unity.com/packages/tools/physics/simple-kepler-orbits-97048