Improved Match Velocity System
Well, I did it. I finally changed the match velocity system to no longer make the player a child of the body, rather, a whole bunch of vector-angular maths was done to move the player with the rotating body whilst keeping them separate.

This should finally eliminate all floating point jitters from the player perspective, and so far, allows planets to be 2 times bigger!